only remove old serverfile if it exists

+++ update.sh
@@ -98,8 +98,8 @@ fi
 # Test internet connectivity and update on success
 wget --spider --quiet https://launcher.mojang.com/v1/objects/1b557e7b033b583cd9f66746b7a9ab1ec1673ced/server.jar
 if [ "$?" != 0 ]; then
-	echo -e "${red}Warning: Unable to connect to Mojang API. Skipping update ...${nocolor}"
-	echo "Warning: Unable to connect to Mojang API. Skipping update ..." >> ${screenlog}
+	echo -e "${red}Warning: Unable to connect to Mojang API. Skipping update...${nocolor}"
+	echo "Warning: Unable to connect to Mojang API. Skipping update..." >> ${screenlog}
 else
 	echo -e "${green}downloading newest server version...${nocolor}"
 	echo "downloading newest server version..." >> ${screenlog}
@@ -110,18 +110,20 @@ else
 		if [ -f "${newserverfile}" ]; then
 			echo -e "${green}Success: updating server.settings for startup with new server version 1.16.5${nocolor}"
 			sed -i "s|${serverfile}|${newserverfile}|g" server.settings
-			rm ${serverfile}
+			if [ -f "${serverfile}" ]; then
+				rm ${serverfile}
+			fi
 		else
-			echo -e "${red}Warning: could not remove old serverfile ${serverfile} because new serverfile ${newserverfile} is missing${nocolor}"
-			echo -e "${yellow}Server will startup with old serverfile ${serverfile}${nocolor}"
+			echo -e "${yellow}Warning: could not remove old serverfile ${serverfile} because new serverfile ${newserverfile} is missing${nocolor}"
+			echo -e "Server will startup with old serverfile ${serverfile}"
 		fi
 fi
 
 # Test internet connectivity and update on success
 wget --spider --quiet https://raw.githubusercontent.com/Simylein/MinecraftServer/master/LICENSE
 if [ "$?" != 0 ]; then
-	echo -e "${red}Warning: Unable to connect to GitHub API. Skipping update ...${nocolor}"
-	echo "Warning: Unable to connect to GitHub API. Skipping update ..." >> ${screenlog}
+	echo -e "${red}Warning: Unable to connect to GitHub API. Skipping update...${nocolor}"
+	echo "Warning: Unable to connect to GitHub API. Skipping update..." >> ${screenlog}
 else
 	echo -e "${green}downloading newest scripts version...${nocolor}"
 	echo "downloading newest scripts version..." >> ${screenlog}
